Add initial solution for day 1, part 2
This solution is incorrect, presumably because it greedily translates
numbers left to right, when it should consider the backwards pass for
digit generation. For example, "eightwo" gives "8wo" at the moment, when
it should give "82", as eight is the first number when read from the
left, but two is the first number when read from the right.